Provide tcsh completion
The new script 'lttng-tcsh_completion' is provided to allow
command-completion for the tcsh shell.

The approach taken by this script is to to re-use the advanced bash
completion script and use its result for tcsh completion.  This is
achieved by running the bash script and outputting the completion
result for tcsh consumption.

This approach is based on a similar script used in the git project.

To use this completion script:

   0) You need tcsh 6.16.00 or newer.
   1) Copy both this file and the bash completion script to ${HOME}.
      You _must_ use the name ${HOME}/.lttng-bash_completion for the
      bash script.
      (e.g. ~/.lttng-tcsh_completion and ~/.lttng-bash_completion).
   2) Add the following line to your .tcshrc/.cshrc:
       source ~/.lttng-tcsh_completion
   3) For completion similar to bash, it is recommended to also
      add the following line to your .tcshrc/.cshrc:
       set autolist=ambiguous
      It will tell tcsh to list the possible completion choices.

Signed-off-by: Marc Khouzam <marc.khouzam@gmail.com>